splits load and forecast requests to 2 lambdas
