+++ Terraform/scrapers/extract_load/Dockerfile
@@ -4,6 +4,6 @@ COPY requirements.txt .
 
 RUN pip install -r requirements.txt --target "${LAMBDA_TASK_ROOT}"
 
-COPY extract.py ${LAMBDA_TASK_ROOT}
+COPY extract_load.py ${LAMBDA_TASK_ROOT}
 
-CMD [ "extract.lambda_handler" ]
+CMD [ "extract_load.lambda_handler" ]
