+++ Terraform/scrapers/extract_load/requirements.txt
@@ -1,0 +1,3 @@
+httpx==0.23.3
+pandas==1.5.2
+pg8000==1.29.4
